// src/api/generated/model/getDrawingResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { GetDrawingResponseAppState } from './getDrawingResponseAppState';
import type { GetDrawingResponseFiles } from './getDrawingResponseFiles';
import type { Time } from './time';

export interface GetDrawingResponse {
  app_state?: GetDrawingResponseAppState;
  created_at?: Time;
  created_by?: string;
  document_id?: string;
  elements?: unknown[];
  files?: GetDrawingResponseFiles;
  icon?: string;
  id?: string;
  name?: string;
  space_id?: string;
  thumbnail?: string;
  updated_at?: Time;
}